protect the list of Peers with a RWMutex
